Combined the initial scraper and updater
Running generate_coins.py will create new csv files for coins that arent included in your coins folder and will update coins that are already there. The dates for coins was changed to y/m/d to fit an incremental/decremental order of numbers. All old code was moved to the legacy folder.